+++ src/components/brukerfeilmelding/brukerfeilmelding.tsx
@@ -4,7 +4,7 @@ import {Brukerfeilmeldinger} from '../../model-interfaces';
 
 export function Brukerfeilmelding(props: Brukerfeilmeldinger) {
     return (
-        <Alert variant={props.variant} size={props.size} inline={props.inline}>
+        <Alert variant={props.variant} size={props.size} inline={true} aria-labelledby={props.text} tabIndex={0}>
             {props.text}
         </Alert>
     );
